new_project/requirements.txt

python-decouple==3.8
Pillow==11.0.0
psycopg2==2.9.5

django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
asgiref==3.8.1
sqlparse==0.5.3
django-autoslug==1.9.9
django-cities-light==3.10.1
progressbar2==4.5.0
python-utils==3.9.1
pytz==2024.2
typing_extensions==4.12.2
unidecode==1.3.8
django-geoposition==0.3.0

djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.10.1

dj-rest-auth==7.0.0

certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
urllib3==2.3.0